Bug Fix: https://app.asana.com/0/579234711066628/605152164458473
Location was not being passed when saving from the settings page.
I didn't remove the location option because I wasn't 100% sure it isn't being used elsewhere.

As it stands now, every one of the fields is checked to see if they exist, if yes, the value will be used else an empty value will be set