Updating "All Jobs" permalink
Previously the "All Jobs" permalink was being built via `<?php echo site_url() . '/' . $post->post_name; ?>`. By appending the `post_name` to the `site_url`, we are just adding the page's slug. This means we're generating the wrong URL for sub-pages. To always get the full URL to the page, we must use `get_permalink()` which is the preferred WordPress way for retrieving a page's URL.